fix qwen vl quantization_func and find_weight_file

+++ tllm/__init__.py
@@ -34,7 +34,7 @@ else:
     import mlx.core as mx
     import numpy as np
 
-    DTYPE = mx.bfloat16
+    DTYPE = mx.float16
     CONVERT_DTYPE = mx.float16
     DES_DTYPE = np.float16
     DEVICE = None
